VMO-4322 add test value & move the `extras` slots
