+++ src/components/interaction-designer/block-types/ConsoleIO_ReadBlock.vue
@@ -13,6 +13,8 @@
         :block="block" />
       <block-name-editor :block="block" />
 
+      <slot name="extras" />
+
       <hr>
 
       <!--Specific config-->
@@ -46,8 +48,6 @@
         :block="block"
         :has-exit-per-choice="false"
         @branchingTypeChangedToUnified="handleBranchingTypeChangedToUnified({block})" />
-
-      <slot name="extras" />
 
       <categorization :block="block" />
