+++ src/components/interaction-designer/block-types/Core_RunFlowBlock.vue
@@ -12,6 +12,8 @@
         :block="block" />
       <block-name-editor :block="block" />
 
+      <slot name="extras" />
+
       <validation-message
         #input-control="{ isValid }"
         :message-key="`block/${block.uuid}/config/flow_id`">
@@ -40,8 +42,6 @@
             :valid-state="isValid" />
         </div>
       </validation-message>
-
-      <slot name="extras" />
 
       <hr>
