+++ src/store/flow/block-types/Core_LogBlockStore.ts
@@ -1,6 +1,6 @@
 import {ActionTree, GetterTree, MutationTree} from 'vuex'
 import {IRootState} from '@/store'
-import {IBlock, IBlockExit} from '@floip/flow-runner'
+import {IBlockExit} from '@floip/flow-runner'
 import {IdGeneratorUuidV4} from '@floip/flow-runner/dist/domain/IdGeneratorUuidV4'
 import {ILogBlock} from '@floip/flow-runner/src/model/block/ILogBlock'
 import {defaultsDeep} from 'lodash'
@@ -41,13 +41,6 @@ export const actions: ActionTree<IFlowsState, IRootState> = {
       tags: [],
       vendor_metadata: {},
     })
-  },
-
-  handleBranchingTypeChangedToUnified({dispatch}, {block}: {block: IBlock}) {
-    dispatch('flow/block_convertExitFormationToUnified', {
-      blockId: block.uuid,
-      test: 'block.value = true',
-    }, {root: true})
   },
 
 }
